pyproject.toml
[project]
name = "agentic-ai-defense-data-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.2",
    "instructor>=1.14.4",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "numpy>=2.2.6",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.2.0",
    "xlsxwriter>=3.2.9",
]
